requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
feedparser
wikipedia
python-dotenv
openai
langchain-google-genai
yfinance
pandas
